Updated library to abstract out the render function and have it's own tests. npm test updated to be recursive.
