+++ package.json
@@ -4,7 +4,7 @@
   "description": "Route for a standardized health-check",
   "main": "index.js",
   "scripts": {
-    "test": "node_modules/.bin/mocha"
+    "test": "node_modules/.bin/mocha \"./{,!(node_modules)/**}/*.test.js\" --recursive"
   },
   "keywords": [
     "health-check",
